Refactor code to better match typical flask project structure
