--- engine/__init__.py
@@ -1,1 +1,0 @@
-from engine.main import Engine
