--- engine/dependency/__init__.py
@@ -1,1 +1,0 @@
-from engine.dependency.main import DependencyGraphBuilder
