--- engine/models/__init__.py
@@ -1,3 +1,0 @@
-from engine.models import train, predict
-
-SAVE_PATH = "models"
