--- engine/models/ngram/__init__.py
@@ -1,1 +1,0 @@
-from engine.models.ngram.model import NGramClassifier
