--- engine/models/rnn/__init__.py
@@ -1,2 +1,0 @@
-from engine.models.rnn.model import RNN
-from engine.models.rnn.model import RNNClassifier
