--- engine/structure/__init__.py
@@ -1,3 +1,0 @@
-from engine.structure.graph import Graph
-from engine.structure.table import Table
-from engine.structure.column import Column
