+++ pagai/engine/dependency/main.py
@@ -2,8 +2,13 @@ from tqdm import tqdm
 import os
 import re
 
-from engine.structure import Graph
-from queries.postgres import get_table_names, get_column_names, get_table, get_column
+from pagai.engine.structure import Graph
+from pagai.services.postgres import (
+    get_table_names,
+    get_column_names,
+    get_table,
+    get_column,
+)
 
 
 EXCLUDE_COLUMNS = ["id", "row_id"]
