+++ pagai/engine/models/__init__.py
@@ -1,0 +1,3 @@
+from pagai.engine.models import train, predict
+
+SAVE_PATH = "pickle"
