+++ pagai/engine/structure/__init__.py
@@ -1,0 +1,3 @@
+from pagai.engine.structure.graph import Graph
+from pagai.engine.structure.table import Table
+from pagai.engine.structure.column import Column
