+++ pagai/services/postgres/main.py
@@ -221,6 +221,10 @@ def fetch_columns(datasets, dataset_size, connection=None):
     print(len(datasets), "datasets")
     for i, dataset in enumerate(datasets):
         print(i, "/", len(datasets))
+
+        if i > 10:
+            continue
+
         table_column_name, nb_datasets = dataset
         table, column = table_column_name.split(".")
